Move format logic to private method
A new method will be created in the future to format with
justify. To better accomodate the logic for this, the
formatting logic moved to a private method which returns
a list of already formatted lines, with no justify, so that
this method can be called by both format-with-no-justity
and format-with-justify methods.